requirements-style.txt
black==23.11.0
django-stubs==4.2.7
djlint==1.34.0
flake8==6.1.0
flake8-commas==2.1.0
flake8-docstrings==1.7.0
flake8-print==5.0.0
flake8-quotes==3.3.2
isort==5.12.0
mypy==1.7.1
